Moved API command "GET version" to "GET qora/version"
