+++ Qora/src/api/ApiService.java
@@ -34,7 +34,6 @@ public class ApiService {
         s.add(ArbitraryTransactionsResource.class);
         s.add(NamePaymentResource.class);
 		s.add(ATResource.class);
-		s.add(VersionResource.class);
 		s.add(BlogPostResource.class);
 		s.add(BlogResource.class);
 		s.add(CalcFeeResource.class);
