+++ Qora/src/api/QoraResource.java
@@ -7,9 +7,12 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 
+import org.json.simple.JSONObject;
+
 import controller.Controller;
 import settings.Settings;
 import utils.APIUtils;
+import utils.BuildTime;
 
 @Path("qora")
 @Produces(MediaType.APPLICATION_JSON)
@@ -58,5 +61,20 @@ public class QoraResource
 	public String getSettings() 
 	{ 
 		return Settings.getInstance().Dump().toJSONString(); 
+	}
+	
+	@SuppressWarnings("unchecked")
+	@GET
+	@Path("/version")
+	public String getVersion()
+	{
+		JSONObject jsonObject = new JSONObject();
+		
+		jsonObject.put("version", Controller.getInstance().getVersion());
+		jsonObject.put("buildDateTime", BuildTime.getBuildDateTimeString());
+		jsonObject.put("buildTimeStamp", BuildTime.getBuildTimestamp());
+	
+
+		return jsonObject.toJSONString();
 	}
 }
